refactor(history): split convertShotRecord + internal helpers out of shothistorystorage.cpp
shothistorystorage.cpp has grown to ~3900 lines covering five concerns
(DB lifecycle, save, load+recompute, serialization, query helpers).
Per the J proposal in openspec/changes/split-shothistorystorage-by-concern/,
this is the first half of the split: the serialization concern plus
the file-static helpers that several TUs need.

Adds:
- shothistorystorage_internal.{h,cpp}: ProfileFrameInfo, AnalysisInputs,
  use12h(), profileFrameInfoFromJson(), prepareAnalysisInputs(). All
  scoped under decenza::storage::detail. Header is internal-only — not
  added to the public include surface.
- shothistorystorage_serialize.cpp: convertShotRecord() and the
  pointsToVariant lambda. Reads helpers via `using` directives at the
  top of the function.

Removes from shothistorystorage.cpp:
- The five file-static helpers (replaced by `using` aliases that point
  at the shared decenza::storage::detail versions).
- The convertShotRecord body (now lives in the serialize TU; main TU
  carries a one-line marker comment at the original location).

CMake: root CMakeLists.txt and tests/CMakeLists.txt updated. The two
MCP test targets that pull shothistorystorage.cpp in directly
(tst_mcptools_write, tst_mcpserver_session) also pick up the two new
TUs via a `replace_all` edit. HISTORY_SOURCES bundles them for the
shared test fixture set.

Pure refactor — no public API change, no behaviour change. Build clean
via Qt Creator MCP. 1811 tests pass, 0 warnings emitted by tests
(the 2 build-time warnings are pre-existing OpenSSL link-version
mismatches in the dev environment).

Queries split (proposal §1.x) is deferred to a follow-up PR for
reviewability — those ~700 lines are independent of this move.
